Renamed package in AndroidManifest and updated R reference in Activities/Fragments/Views files.
