+++ app/src/main/java/com/chess/clock/activities/ClockTimersActivity.java
@@ -1,4 +1,4 @@
-package com.chess.clock.app;
+package com.chess.clock.activities;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -20,6 +20,8 @@ import android.view.View;
 import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.TextView;
+
+import com.chess.clock.R;
 import com.chess.clock.engine.CountDownTimer;
 import com.chess.clock.engine.Stage;
 import com.chess.clock.engine.TimeControlParser;
